pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.14.1",
    "litellm>=1.77.4",
    "requests>=2.32.0",
    "python-dotenv>=1.1.1",
]
